Added new methods to DbHelper to get array lists of Cuisines, Ingredients, MealTypes

+++ app/src/main/java/ca/uottawa/cookingwithgarzon/model/MealType.java
@@ -12,7 +12,7 @@ public class MealType {
         return _id;
     }
 
-    public void set_id(int _id) {
+    public void set_id(long _id) {
         this._id = _id;
     }
